requirements.txt
alembic==1.16.2
aniso8601==10.0.1
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.6.15
click==8.2.1
distlib==0.3.9
filelock==3.18.0
Flask==3.1.1
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.0
packaging==25.0
pipenv==2025.0.3
platformdirs==4.3.8
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
virtualenv==20.31.2
Werkzeug==3.1.3
